refactor(types): extract Footer
