+++ dotcom-rendering/src/types/footer.ts
@@ -1,0 +1,10 @@
+interface FooterLink {
+	text: string;
+	url: string;
+	dataLinkName: string;
+	extraClasses?: string;
+}
+
+export interface FooterType {
+	footerLinks: FooterLink[][];
+}
